Return 0 instead of doing another rt:sysread
Previously, wait_until_read function would execute
another rt:systest_read call and return a count of
errors, when we already asserted above that the list
is empty.

If the wait_until function does not return true, or
times out, or hits the retry limit, the running test
ought to fail anyway without needing an additional
read to fail.

So doing another call is redundant and unnecessary;
we just return 0 as the value if rt:wait_until
completes successfully.