Fix false positive test fails caused by AAE race
It is occasionally possible to see a YZ AAE repair in normal operation
if we happen to be running the exchange FSM while writes are occurring,
and we compare the KV tree (which already has the new data inserted)
with the YZ tree (which doesn't have the data inserted yet).

This ends up triggering an extra call to get_map, which could cause some
earlier parts of this test to fail when it expected 60 calls but saw 61.

To fix this, we can simply disable AAE at the start of the test, and
then enable it later when we get to the part of the test that requires
AAE.