Copy refactored rolling_upgrade from JV's work
This version doesn't muck around with the AAE config for no reason, and
also avoids a bunch of redundant code which is a nice bonus. Existing
tests using this function continue to work fine with the changes.

At some point we may want to get rid of the copy of this function that
lives in yokozuna, so we can avoid having a duplicate function, but
until everything gets merged to the appropriate place we'll have to live
with it.