Simplify get_victim & fix intermittent w1c fails
There's no reason I can find to rule out any particular vnode when
choosing the victim. The way this function is used in this test, it
seems we're trying to make sure we pick a vnode on a different node from
the one we're connecting to, but I removed this check and forced it to
pick a vnode on the same node, and the test still passed fine.

This was also causing occasional issues with the w1c tests, because
there we use n=1; if we skip the first entry in the preflist because
it's on the excluded node, then we will pick a victim vnode that is not
actually being read from, and the test will fail.