Re-add old kv679 test
This test was part of the original kv679 suite, but at the time of 2.1
it was put in it's own branch. Riak Test is made up of passing tests,
and this test (kv679_dataloss_fb.erl) still fails. Adding this back to a
branch off develop in preparation for a fix of this issue.