Added tests for pd value. When using w=3, pw=2 in the attempt to prevent data loss by writing primary nodes only to ensure the write goes to more than one physical node, one ends up rejecting writes in the case of more than one node going down. pd (physical diversity) solves this issue by writing to both primary and fallback nodes, ensuring that the writes are to different physical nodes.
This test demonstrates that of writing to a bucket with pw=2 when 2 nodes
from the preflist are down will be rejected, whereas the same situation with
pd=2 returns a successful write.
Finally, it demonstrates that write to a bucket with a pd value that cannot
be met will be rejected.