Remove `jmx` property from riak.conf on upgrade
With the move from riak_ee to OS riak+repl for 2.2.5, jmx and snmp have
been dropped. Riak failed to start up with the old riak.conf from 2.0.x
or 2.2.N<4 as it contained a property for `jmx` that there is no longer
a schema for. The solution is a hack due to cuttlefish deficiency
described in this issue https://github.com/basho/cuttlefish/issues/176