Add mystery sleeps
The sleeps, for reasons unnknown, make the tests pass consistently.

This is unsatisfactory - and will be tracked through https://github.com/basho/riak_test/issues/1313.

However, initially, for this release, it is a good enough answer to proceed (in that it allows us to re-add the http_security tests in which have shown to be important).