Add test that forces rebuild
It transpires that the `verify_tictac_aae` test does a poor job of actually testing tictca aae rebuilds.  Even if rebuilds fail, the pace of exchanges will repair discrepancies via read repair.

This test explicitly requires the reuild to work, and cannot solve any issues through exchanges (as exchanges are stopped for the test).

Also fills a gap in testing of modified date ranges in aaefold.