Avoid potential race
Within wait_and_validate - especially when testing the addition of Node 4 in OTP 20 - there may be a race condition.

In this condition the nodes may be very briefly in a state whereby they are up and ready, but the pending changes are not yet on the ring.  In this case rt:wait_until_no_pending_changes/1 may pass before pending changes have completed.

In this case, a subsequent change may to bring a node down may hang waiting for pending changes, as the pending changes cannot complete due to the node now being down.

To avoid this, we re-wait for no pending changes at the end of each cycle.