Changes to pass some kv_all tests
riak727 is removed from kv_all upgrade, as the test is a test of an upgrade from 1.4.  It has been stated that for 3.0 upgrades will be supported from 2.9 and 2.9.1 only.

Fix to path in kv679 dataloss to reflect change in riak `make devrel` script which now includes `riak` within the path.

The pb_cipher_suites test has been changed to reflect that protocols before tLS1.2 are no longer supported.  Also, the implementation of honor_cipher_order has switched from patched OTP to mainstream OTP - and as part of this change the default of honor_cipher_order has been switched from false to true.  So now, to honour cipher_order rhis must be explicitly set to true.  This si a change that *must* be recognised in release notes.  It may have been possible to change the code to implement the previous default behaviour, but having an explicit difference between riak and OTP defaults may cause long term confusion.

https://github.com/erlang/otp/blob/OTP-20.3.8.22/lib/ssl/src/ssl_internal.hrl#L133-L135

https://github.com/erlang/otp/edit/maint/lib/ssl/doc/src/ssl.xml#L1034

The riak_rex testwas a test of safe_rpc - that rpc to another node would not throw an exception.  Now, though OTP has changed to stop it from throwing an exception.

https://github.com/erlang/otp/commit/cccf762d93640b707f8e1515604cebcfbe379c04

This means that the safe_rpc try/catch is no longer necessary.  Open question about whether the safe_rpc wrapper should be changed to return the previous error for backward compatability (although nothin other than the test appears to exist in the code to handle this specific error), or whether the try/catch should just be removed