Set target_n_val to at least n_val
See - https://docs.riak.com/riak/kv/latest/configuring/strong-consistency/index.html#setting-the-target-n-val-parameter

Previously ensemble_util would not set the target_n_val, and it would default to 4 when nval is 5 in many riak_ensemble tests.

This setting could mean that a non-quorum minority partition (2 nodes) could contain a majority of peers (3 peers) - and so it would not be possible for all ensembles to prove quorum during such a partition.

Where tests are going to partition the cluster they need to have at least nval + 1 nodes - as without this the target_n_val will not be achievable.

Also stops over-riding backend.  This runs the risk that the tests assume that the backend has the features of the in-memory backend (e.g. non-persisting) - but that isn't documented.  The tests appear to be more stable with bitcask backend.  All tests still at least past intermittently with the in-memory backend.

The test also uses more relaistic defaults.  It increases the super-fast anti-entropy tick, and normalises the vnode_management_timer.  This appeasr to increase stability in tests (at a cost that more sleeps required to give time for completion).