OTP compatability for cipher_suites
Some test wrangling to get pb_cipher_suites to pass on all OTP versions.

Some caveats:

The API is not compatible - we rely on the user knowing what OTP version they're running and interfacing with the riak erlang client the correct way (passing in the right ciphers etc).

Testing of honor_cipher_order doesn't seem to work - may be a cipher compatability issue.  Suspect this is a test issue, but ultimately this is well tested by the OTP team, so do we need to repeat here?  There is always the workaround of submitting just the desired cipher.