Add recognition of 2i stats (#1359)
* Add recognition of 2i stats

* Add new windex stats to test

* Update verify_riak_stats.erl

* Check new stats via statsfold test

Also extends basic_upgrade test to test a slightly less trivial volume of data, and include in a group.

* Ensure queue by repeating stats fetch, and reducing wait

* Intermittent failure - with immediate delete

This test would intermittently fail.  The failure is due to immediate delete not always reaping tombstones (perhaps due to race conditions).  This appears to be happening to random keys.

We now tolerate some failure for immediate delete to reap tombstones, rather than failing the test.  Testing the function of immediate delete is not the purpose of this test.