Update for 3.2 tests
redbug doesn't generate a file if the monitored function does not get called - and so report a count of 0 if the file does not exist.

No syslogger in 3.2 release candidate, so this is dropped from the riak stats test.

It is possible to get timeouts on writes/reads immediately before after partition/healing.  This is not relevant to the tested functionality ... so pause a bit further to try and avoid.

Reminder that admin bucket type creation doesn't necessarily work on OSX.  Supported OS required