Use riak stats to confirm
The test has two functions:

Issue 1759 (KV).  Confirm that through all the leaves and rejoins false AAE does not kick in because old state is left lingering.

Issue 994 (Core).  Confirm the controlling read-repair to fallbacks works as expected.  Nodes that are configured not to repair fallbacks don't, but still repair primaries.